Pack of RD lifetimes fixes + auto-binding for task results

+++ rd-cpp/src/rd_core_cpp/src/main/lifetime/SequentialLifetimes.cpp
@@ -2,32 +2,32 @@
 
 namespace rd
 {
-SequentialLifetimes::SequentialLifetimes(Lifetime parent_lifetime) : parent_lifetime(std::move(parent_lifetime))
+SequentialLifetimes::SequentialLifetimes(const Lifetime& parent_lifetime) : parent_lifetime(parent_lifetime)
 {
-	this->parent_lifetime->add_action([this] { set_current_lifetime(LifetimeDefinition::get_shared_eternal()); });
 }
 
 Lifetime SequentialLifetimes::next()
 {
-	std::shared_ptr<LifetimeDefinition> new_def = std::make_shared<LifetimeDefinition>(parent_lifetime);
-	set_current_lifetime(new_def);
-	return current_def->lifetime;
+	Lifetime new_lifetime = parent_lifetime.create_nested();
+	set_current_lifetime(new_lifetime);
+	return new_lifetime;
 }
 
 void SequentialLifetimes::terminate_current()
 {
-	set_current_lifetime(LifetimeDefinition::get_shared_eternal());
+	set_current_lifetime(Lifetime::Terminated());
 }
 
 bool SequentialLifetimes::is_terminated() const
 {
-	return current_def->is_eternal() || current_def->is_terminated();
+	return current_lifetime->is_terminated();
 }
 
-void SequentialLifetimes::set_current_lifetime(std::shared_ptr<LifetimeDefinition> new_def)
+void SequentialLifetimes::set_current_lifetime(const Lifetime& lifetime)
 {
-	std::shared_ptr<LifetimeDefinition> prev = current_def;
-	current_def = new_def;
-	prev->terminate();
+	const Lifetime prev = current_lifetime;
+	current_lifetime = lifetime;
+	if (!prev->is_terminated())
+		prev->terminate();
 }
 }	 // namespace rd
